requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
django-filter==23.3
dj-database-url==2.1.0
Pillow==10.1.0
pydantic==2.4.2
typing-extensions==4.8.0